refactored reviews, now on to refactoring likes
